refactor: anti-monolith modularization of Setup, add universal Minecraft and CI/CD workflow
